Move job data to file
