+++ UI/ConfigUI.cs
@@ -1,8 +1,13 @@
-﻿using ImGuiNET;
+﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+using ImGuiNET;
 
-namespace DeepDungeonDex
+namespace DeepDungeonDex.UI
 {
     public class ConfigUI
     {
